# [mempool] Expire transactions early, to avoid having txn's in block that will expire before getting on-chain