Before (committed/s drops to half):

```
wanted/s     | submitted/s  | committed/s  | expired/s    | rejected/s   | chain txn/s  | latency      | p50 lat      | p90 lat      | p99 lat      | actual dur  
7000         | 7001         | 6198         | 803          | 0            | 6091         | 34116        | 32000        | 60600        | 65800        | 360         
8000         | 7993         | 3872         | 4120         | 0            | 6166         | 51079        | 55800        | 70300        | 73600        | 360         
10000        | 9993         | 3390         | 6602         | 1698         | 6250         | 58066        | 65800        | 71800        | 75300        | 360         
12000        | 11991        | 3398         | 8592         | 2576         | 6166         | 63447        | 64600        | 70000        | 73600        | 360      

```
After (very low degradation in committed/s, much more gracefully handling the situation):
```

wanted/s     | submitted/s  | committed/s  | expired/s    | rejected/s   | chain txn/s  | latency      | p50 lat      | p90 lat      | p99 lat      | actual dur  
8000         | 8004         | 6288         | 1715         | 0            | 5986         | 44779        | 47900        | 50700        | 52200        | 432         
10000        | 9999         | 5658         | 4340         | 0            | 5708         | 49995        | 49800        | 52800        | 54900        | 432         
12000        | 11993        | 5523         | 6469         | 0            | 5666         | 48895        | 48900        | 52000        | 54400        | 432  
```       



### Description

### Test Plan
<!-- Please provide us with clear details for verifying that your changes work. -->

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/aptos-labs/aptos-core/4188)
<!-- Reviewable:end -->