@igor-aptos 2022-09-14
@banool , @gregnazario  - I am adding a new error code here. Is that a breaking change or not? should I add a new code to enum, but not yet use it, and use an existing InternalError one instead?