@igor-aptos 2022-09-14
committed/s is how many transactions were successfully committed

chain/s is how many transactions got onto blockchain. previously, half of the transaction would get onto blockchain, but have their expiration time after the block time, and so be rejected. which is what we are trying to fix here.