@igor-aptos 2022-09-14
> So, you're making mempool expire transactions before they expire. I'm not sure that makes any sense, because now you're setting a validity range of the time set but the user, and + 20 seconds from now.

Yes. We select transactions with expiry times as close to now as possible. But then there is a delay until they get committed on-chain, and basically if their expiry time is too close - they will be rejected. 

under load p50 is 10s, p90 is 20s. Even without load, it is in seconds. 
This PR makes the threshold be constant. 
Alternative would be to use metric on how much time it takes from taking from mempool, to getting on-chain, and expire based on that, but then we wouldn't really be able to notify users on submission why we expired the transaction.